Split structure commands into their own file
